+++ sphinx/application/common/wrappers/wrapper-common/src/main/java/chat/sphinx/wrapper_common/HideBalance.kt
@@ -1,0 +1,11 @@
+package chat.sphinx.wrapper_common
+
+object HideBalance {
+
+    const val ENABLED = 1
+    const val DISABLED = 0
+
+    const val HIDE_BALANCE_SHARED_PREFERENCES = "general_settings"
+    const val HIDE_BALANCE_ENABLED_KEY = "hide-balance-enabled"
+
+}